feat: implement uksi and bold data importing
Two cli commands, one requires a BOLD data package the other a UKSI DwC-A. These both require tests and docs but work for the moment while I feel out what data we actually need.
Additionally, this commit comes with the basic models as well. These haven't been finalised for a first version so no migrations exist yet - this just makes it easier to dev against and avoids creating hundreds of migration versions while I add, remove, and change model defs during initial dev.
Eventually, the UKSI data should come from an API, either the TS provided species-dictionary or the Data Portal (it'll probably be the data portal which itself will get the data from the TS API). However, currently the API istoo slow to use and is being migrated in k8s. Once that migration is complete we should look at switching the import code so that it uses the API instead of a DwC-A, or indeed, the web app just uses the API directly (if whichever API we use is fast enough for this, and we don't end up adding references from the BOLD data to the UKSI data in the model). This raises another point, there are currently no database references between the BOLD data and the UKSI data (the taxonomic information on each exists separately). This is partly a speed choice, we don't need it right now and it's not fully clear yet what we might need, and also a pragmatic choice - matching taxonomies is hardish and we should only do it if necessary. I think we might be able to do it at query time with decent performance which would also allow us flexibility with changing the matching algorithm and not requiring a full re-import of the data from BOLD/UKSI but we'll see if that proves true or not.