fix: remove taxon nodes with missing parent nodes
The current data coming from the NBN API for UKSI includes two taxa (NHMSYS0001485599 and NHMSYS0001485598) which have NHMSYS0001485597 as their parent but this parent ID doesn't exist on NBN anywhere. This broke the existing code for a few reasons, but ultimately necessitated a fix because it will not work in the database with the parent ID foreign key constraint.