fix: remove keepalive from app root
For the moment at least. It would be nice to have this because it makes the app more efficient and could be used to maintain state when navigating between the various tabs but at the moment for the data page it just makes things not work in a sensible way (if you load a species and then navigate to home and then back to the data page you get the full tree loaded down to the species but not the species page itself). Really we should implement pinia and store some state to make this easier to manage.