fix: show loading spinner on species view always when required
Prior to this, the spinner only showed on first load as it was part of the suspense. This makes it part of the component itself so it can manage it. This also reimplements how reloads are managed on the species view component to be much better and cleaner. Need to do the sidebar now too but it works for now so future task...