fix: collect specimens correctly for BIN aggregating endpoints
Prior to this change, the aggregated BINs were filtered by the taxon name and therefore only aggregated specimens with the requested taxon name. This is wrong according to Ben/the spec so this is corrected here. What we actually need to do is find the BINs the given taxon name appears in and then aggregate over all the specimens in those BINs regardless of whether they have the same taxon name as the original request. This is now corrected in the 2 endpoints exposing this kind of data.

Closes: #5